chore(wallet): move formatDateNow to util file
